services/news-archiver/requirements.txt
# services/news-archiver/requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
langchain-core>=0.1.0
langchain-qdrant>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.0.1
